Map correct mediatype for transmission attachments
