+++ src/Altinn.Correspondence.Integrations/Dialogporten/Mappers/CreateDialogTransmissionMapper.cs
@@ -97,6 +97,9 @@ namespace Altinn.Correspondence.Integrations.Dialogporten.Mappers
             var baseTimestamp = DateTimeOffset.UtcNow;
             return correspondence.Content?.Attachments.Select((correspondenceAttachment, index) =>
             {
+                var attachmentFileName = correspondenceAttachment?.Attachment?.FileName;
+                var mediaType = DialogportenAttachmentMediaTypeMapper.GetDialogportenAttachmentMediaTypeForFileName(attachmentFileName);
+
                 var transmissionAttachment = new TransmissionAttachment
                 {
                     Id = Guid.CreateVersion7(baseTimestamp.AddMilliseconds(index)).ToString(),
@@ -113,7 +116,7 @@ namespace Altinn.Correspondence.Integrations.Dialogporten.Mappers
                         new TransmissionUrl
                         {
                             ConsumerType = "Gui",
-                            MediaType = "application/vnd.dialogporten.frontchannelembed-url;type=text/markdown",
+                            MediaType = mediaType,
                             Url = GetDownloadAttachmentEndpoint(baseUrl, correspondence.Id, correspondenceAttachment.AttachmentId)
                         }
                     }
