+++ src/Altinn.Correspondence.Integrations/Dialogporten/Models/CreateTransmissionRequest.cs
@@ -122,7 +122,8 @@ public class TransmissionAttachment
         public string Url { get; set; }
 
         [JsonPropertyName("mediaType")]
-        public string MediaType { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
+        public string? MediaType { get; set; }
 
         [JsonPropertyName("consumerType")]
         public string ConsumerType { get; set; }
